LayoutTests/http/tests/history/cross-origin-replace-history-object-expected.txt

ALERT: Parent window's history object before attempt to clear: [object History]
ALERT: Attempting to clear parent window's history object:
CONSOLE MESSAGE: line 21: SecurityError: Blocked a frame with origin "http://localhost:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
ALERT: Parent window's history object after attempt to clear: [object History]
